refactor(index-generation): split certification internals
